+++ packages/devtools_app/lib/src/screens/memory/panes/diff/widgets/snapshot_control_pane.dart
@@ -12,7 +12,7 @@ import '../../../../../shared/memory/simple_items.dart';
 import '../../../../../shared/primitives/byte_utils.dart';
 import '../../../shared/primitives/simple_elements.dart';
 import '../controller/diff_pane_controller.dart';
-import '../controller/item_controller.dart';
+import '../controller/snapshot_item.dart';
 
 class SnapshotControlPane extends StatelessWidget {
   const SnapshotControlPane({super.key, required this.controller});
@@ -76,8 +76,8 @@ class _DiffDropdown extends StatelessWidget {
 
   List<DropdownMenuItem<SnapshotDataItem>> items() =>
       controller.core.snapshots.value
-          .where((item) => item.hasData)
-          .cast<SnapshotDataItem>()
+          .whereType<SnapshotDataItem>()
+          .where((item) => item.isProcessed)
           .map(
         (item) {
           return DropdownMenuItem<SnapshotDataItem>(